+++ packages/ragbits-document-search/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Add LLM reranker (#513)
+- fix: fix passing score in DocumentSearch.search (#545)
 
 ## 0.16.0 (2025-04-29)