+++ packages/ragbits-document-search/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+## 0.17.0 (2025-05-06)
+
+### Changed
+
+- ragbits-core updated to version v0.17.0
+
 - Add optional score override on reranking step (#544)
 - Add score threshold to reranker options (#544)
 - Add LLM reranker (#513)